+++ pages/api/deleteProject.js
@@ -1,15 +1,17 @@
 import { initEndPoint, query } from '/lib/api-helpers'
 import { checkUserIsOwner } from '/lib/api-database-helpers'
 
-export default async function handler(request, response) {
-  await initEndPoint(request, response, async (userId, { projectId }) => {
-    if (!await checkUserIsOwner(projectId, userId)) {
-      return { error: 'userDoesNotHavePermissionToDeleteProject' }
-    }
+export default initEndPoint(async (userId, { projectId }) => {
+  if (userId == undefined) {
+    return { error: 'userNotLoggedIn' }
+  }
 
-    await query(`
-      DELETE FROM Project WHERE ID_Project = ?`,
-      [projectId])
-    return { ok: {} }
-  })
-}
+  if (!await checkUserIsOwner(projectId, userId)) {
+    return { error: 'userDoesNotHavePermissionToDeleteProject' }
+  }
+
+  await query(`
+    DELETE FROM Project WHERE ID_Project = ?`,
+    [projectId])
+  return { ok: {} }
+})
