+++ pages/api/deleteTask.js
@@ -1,20 +1,22 @@
 import { initEndPoint, query } from '/lib/api-helpers'
 import { checkUserIsOwner, checkUserIsAdmin, getTaskProject } from '/lib/api-database-helpers'
 
-export default async function handler(request, response) {
-  await initEndPoint(request, response, async (userId, { taskId }) => {
-    const projectId = await getTaskProject(taskId)
-    if (projectId == null) {
-      return { error: 'taskIsNotExist' }
-    }
+export default initEndPoint(async (userId, { taskId }) => {
+  if (userId == undefined) {
+    return { error: 'userNotLoggedIn' }
+  }
 
-    if (!(await checkUserIsOwner(projectId, userId) || await checkUserIsAdmin(projectId, userId))) {
-      return { error: 'userDoesNotHavePermissionToDeleteTasks' }
-    }
+  const projectId = await getTaskProject(taskId)
+  if (projectId == null) {
+    return { error: 'taskIsNotExist' }
+  }
 
-    await query(`
-      DELETE FROM Task WHERE ID_Task = ?;`,
-      [taskId])
-    return { ok: {} }
-  })
-}
+  if (!(await checkUserIsOwner(projectId, userId) || await checkUserIsAdmin(projectId, userId))) {
+    return { error: 'userDoesNotHavePermissionToDeleteTasks' }
+  }
+
+  await query(`
+    DELETE FROM Task WHERE ID_Task = ?;`,
+    [taskId])
+  return { ok: {} }
+})
